Move tests into one
